refactor(tui): separate terminal browser responsibilities
